fTest()/fcmpx: third parameter (=exponent difference) were loaded into FPU, which is wrong.
The exponent difference is pass to the back ends as memory operator of type [S]DWORD. Currently the back ends forwards this value to the library function fpu_zero_test using the INVOKE directive. The macro fcmpx did load this value in to the FPU before calling the back end, which is wrong because the back end didn't expect that and left the value on FPU stack when leaving.